Minor fix for gcc builidng problem.
Added explicit type conversion in initilization list.

BUG=None
TEST=Built using gcc 4.7.


Review URL: https://chromiumcodereview.appspot.com/11193050

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@162945 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 4d22414663c986c3af4a7af5ce9a062bb1b405b4

+++ sandbox/linux/seccomp-bpf/verifier.cc
@@ -27,7 +27,8 @@ bool Verifier::VerifyBPF(const std::vector<struct sock_filter>& program,
     // indicates either i386 or x86-64; and a set bit 30 indicates x32. And
     // unless we pay attention to setting this bit correctly, an early check in
     // our BPF program will make us fail with a misleading error code.
-    struct arch_seccomp_data data = { sysnum, SECCOMP_ARCH };
+    struct arch_seccomp_data data = { static_cast<int>(sysnum),
+                                      static_cast<uint32_t>(SECCOMP_ARCH) };
 #if defined(__i386__) || defined(__x86_64__)
 #if defined(__x86_64__) && defined(__ILP32__)
     if (!(sysnum & 0x40000000u)) {
